tests: refactor a bit
